Readme update, battlelib component rename for accuracy, splitting config and token (closes #23)
