+++ .gitignore
@@ -20,3 +20,4 @@ node_modules/
 .config
 
 battlecache.json
+.token.json
