+++ battlelib/battlecachedao.js
