+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "beatbattlebot",
-  "version": "1.0.5",
+  "version": "1.0.6",
   "description": "Discord.js bot for organizing and handling beat battle submissions on music servers",
   "main": "bot.js",
   "scripts": {
